Add colour contrast checking library
In order to be as accessible as possible we should make sure that
foreground and background colours are chosen such that they meet WebAIM
principles and rules. This includes in things like loggers etc., where
we might not normally think to include them.

This adds a library to handle this and flag up is something isn't
accessible.