refactor(getCookie): retrieve memoized directly
There’s no need to check if the value exists before
retrieving it, as the default behaviour of `get` on a Map
is to return undefined if the key is missing. If the key
exists, return it directly.

This allows us to reach true 100% coverage,
as the missing code path was unreachable.