chore: change dependabot frequency
make dependabot run once a month, and do not limit the amount of PRs it opens.

we're going to try manually bumping deps weekly and the theory is that dependabot will have v little to do when it's time to run again. the monthly run is intended to be shouty in case we've not got round to it.